requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
flask-swagger-ui==4.11.1
psycopg2-binary==2.9.9
cloudinary==1.36.0
python-dotenv==1.0.0
Werkzeug==3.0.1
